add logic file Relates #43
